feat: add prompt example
