+++ spring-ai-alibaba-prompt-example/README.md
@@ -1,0 +1,3 @@
+# Spring AI Alibaba Prompt Example
+
+此示例将演示如何使用 Spring AI 提供的 Prompt 模板配置功能。
