--- spring-ai-alibaba-prompt-example/prompt-example/src/main/java/com/alibaba/cloud/ai/example/prompt/PromptExampleApplication.java
@@ -1,13 +1,0 @@
-package com.alibaba.cloud.ai.example.prompt;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PromptExampleApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PromptExampleApplication.class, args);
-	}
-
-}
