--- spring-ai-alibaba-prompt-example/prompt-example/src/main/java/com/alibaba/cloud/ai/example/prompt/roles/Completion.java
@@ -1,15 +1,0 @@
-package com.alibaba.cloud.ai.example.prompt.roles;
-
-public class Completion {
-
-	private final String completion;
-
-	public Completion(String completion) {
-		this.completion = completion;
-	}
-
-	public String getCompletion() {
-		return completion;
-	}
-
-}
