--- spring-ai-alibaba-prompt-example/prompt-example/src/main/resources/application.yml
@@ -1,10 +1,0 @@
-spring:
-  application:
-    name: prompt-example
-  ai:
-    dashscope:
-      api-key: ${AI_DASHSCOPE_API_KEY}
-  nacos:
-    config:
-      enabled: true
-      server-addr: localhost:8848
