+++ spring-ai-alibaba-prompt-example/src/main/java/com/alibaba/cloud/ai/example/prompt/controller/RoleController.java
@@ -1,4 +1,21 @@
-package com.alibaba.cloud.ai.example.prompt.roles;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.cloud.ai.example.prompt.controller;
 
 import java.util.List;
 import java.util.Map;
@@ -13,10 +30,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/example/ai")
 public class RoleController {
 
 	private final ChatClient chatClient;
@@ -29,16 +48,24 @@ public class RoleController {
 		this.chatClient = builder.build();
 	}
 
-	@GetMapping("/ai/roles")
-	public AssistantMessage generate(@RequestParam(value = "message",
+	@GetMapping("/roles")
+	public AssistantMessage generate(
+			@RequestParam(value = "message",
 			defaultValue = "Tell me about three famous pirates from the Golden Age of Piracy and why they did.  Write at least a sentence for each pirate.") String message,
 			@RequestParam(value = "name", defaultValue = "Bob") String name,
-			@RequestParam(value = "voice", defaultValue = "pirate") String voice) {
+			@RequestParam(value = "voice", defaultValue = "pirate") String voice
+	) {
+
 		UserMessage userMessage = new UserMessage(message);
+
 		SystemPromptTemplate systemPromptTemplate = new SystemPromptTemplate(systemResource);
 		Message systemMessage = systemPromptTemplate.createMessage(Map.of("name", name, "voice", voice));
-		Prompt prompt = new Prompt(List.of(userMessage, systemMessage));
-		return chatClient.prompt(prompt).call().chatResponse().getResult().getOutput();
+
+		return chatClient.prompt(new Prompt(List.of(userMessage, systemMessage)))
+				.call()
+				.chatResponse()
+				.getResult()
+				.getOutput();
 	}
 
 }
