+++ spring-ai-alibaba-prompt-example/src/main/java/com/alibaba/cloud/ai/example/prompt/controller/StuffController.java
@@ -1,0 +1,77 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.cloud.ai.example.prompt.controller;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import com.alibaba.cloud.ai.example.prompt.entity.Completion;
+
+import org.springframework.ai.chat.client.ChatClient;
+import org.springframework.ai.chat.prompt.PromptTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.Resource;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/example/ai")
+public class StuffController {
+
+	private final ChatClient chatClient;
+
+	@Value("classpath:/docs/wikipedia-curling.md")
+	private Resource docsToStuffResource;
+
+	@Value("classpath:/prompts/qa-prompt.st")
+	private Resource qaPromptResource;
+
+	@Autowired
+	public StuffController(ChatClient.Builder builder) {
+		this.chatClient = builder.build();
+	}
+
+	@GetMapping(value = "/stuff")
+	public Completion completion(
+			@RequestParam(value = "message",
+			defaultValue = "Which athletes won the mixed doubles gold medal in curling at the 2022 Winter Olympics?'") String message,
+			@RequestParam(value = "stuffit", defaultValue = "false") boolean stuffit
+	) {
+
+		PromptTemplate promptTemplate = new PromptTemplate(qaPromptResource);
+
+		Map<String, Object> map = new HashMap<>();
+		map.put("question", message);
+		if (stuffit) {
+			map.put("context", docsToStuffResource);
+		}
+		else {
+			map.put("context", "");
+		}
+
+		return new Completion(
+				chatClient.prompt(promptTemplate.create(map))
+						.call()
+						.content()
+		);
+	}
+
+}
