+++ spring-ai-alibaba-prompt-example/src/main/resources/data/bikes.json
