+++ spring-ai-alibaba-prompt-example/src/main/resources/docs/wikipedia-curling.md
