+++ spring-ai-alibaba-prompt-example/src/main/resources/prompts/joke-prompt.st
