+++ spring-ai-alibaba-prompt-example/src/main/resources/prompts/qa-prompt.st
