+++ spring-ai-alibaba-prompt-example/src/main/resources/prompts/system-message.st
